WD-24: big component refactor
